requirements.txt
# Discord-Claude Bot Framework Requirements
# Python 3.9+

# Discord API
discord.py>=2.3.0

# Anthropic Claude API
anthropic>=0.45.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client
aiohttp>=3.9.0

# Image processing
Pillow>=10.0.0

# Configuration
pyyaml>=6.0.0

# SQLite async support
aiosqlite>=0.19.0
